+++ src/main/java/br/com/fintech/API/wallet/model/dto/WalletResponseDTO.java
@@ -1,9 +1,8 @@
 package br.com.fintech.API.wallet.model.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record WalletResponseDTO(
-        BigDecimal balance,
+        Double balance,
         List<OperationDTO> operations
 ) {}
